Tidy up amp-story-page block components
